initial

+++ cifar100_ppio.py
@@ -93,7 +93,7 @@ def main():
     model = ResNet50Classifier()
     # trainer = Trainer(accelerator='npu', devices='0,1', max_epochs=5, strategy='deepspeed', precision=16)
     # trainer = Trainer(accelerator='gpu', devices=1, max_epochs=5, precision=16, strategy='deepspeed')
-    trainer = Trainer(accelerator='gpu', devices=1, max_epochs=5, precision=16)
+    trainer = Trainer(accelerator='gpu', devices=1, max_epochs=5)
     trainer.fit(model, datamodule=data_module)
     trainer.test(model, datamodule=data_module)
